feat: allow weekend assignment gaps and reuse edit dialogs
