+++ frontend/src/features/localization/translations.ts
@@ -117,7 +117,7 @@ const translations = {
       schoolMissing: 'Keine Schule hinterlegt.',
       notes: 'Notizen',
       internships: 'Praktikumszeiträume',
-      internshipsDescription: 'Ein Zeitraum muss lückenlos mit Teamzuweisungen abgedeckt sein.',
+      internshipsDescription: 'Alle Werktage eines Zeitraums müssen durch Teamzuweisungen abgedeckt sein. Wochenendlücken sind erlaubt.',
       addInternship: 'Zeitraum',
       noInternshipsYet: 'Noch kein Zeitraum angelegt.',
       internshipTitle: 'Zeitraum {index}',
@@ -126,7 +126,7 @@ const translations = {
       endDate: 'Enddatum',
       internshipNote: 'Notiz zum Zeitraum',
       assignments: 'Teamzuweisungen',
-      assignmentsDescription: 'Ein Wechsel innerhalb einer Woche ist über das Datumsintervall möglich.',
+      assignmentsDescription: 'Ein Wechsel innerhalb einer Woche ist über das Datumsintervall möglich. Wochenenden können ohne Zuweisung bleiben.',
       addAssignment: 'Zuweisung',
       team: 'Team',
       teamPlaceholder: 'Team wählen',
@@ -451,7 +451,7 @@ const translations = {
       schoolMissing: 'No school stored.',
       notes: 'Notes',
       internships: 'Internship periods',
-      internshipsDescription: 'Each period must be fully covered by team assignments without gaps.',
+      internshipsDescription: 'Each weekday in a period must be covered by team assignments. Weekend gaps are allowed.',
       addInternship: 'Period',
       noInternshipsYet: 'No period created yet.',
       internshipTitle: 'Period {index}',
@@ -460,7 +460,7 @@ const translations = {
       endDate: 'End date',
       internshipNote: 'Period note',
       assignments: 'Team assignments',
-      assignmentsDescription: 'Switches within a week are possible through the date range.',
+      assignmentsDescription: 'Switches within a week are possible through the date range. Weekend gaps can stay unassigned.',
       addAssignment: 'Assignment',
       team: 'Team',
       teamPlaceholder: 'Select team',
